backend/requirements.txt
# Add CORS support to Flask backend
# from flask_cors import CORS

flask
flask-cors
sqlalchemy
python-dotenv
PyJWT[crypto]
google-genai
gunicorn
psycopg2-binary